legend/projects/godata/pyproject.toml
[project]
name = "godata"
version = "0.1.0"
description = "八卦纳甲映射系统 - 传统文化数据工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psutil>=7.0.0",
    "rich>=14.0.0",
    "sxtwl>=2.0.7",
    "fastapi>=0.104.0",
    "pydantic>=2.4.2",
    "uvicorn>=0.23.2",
    "idp-framework @ file:///workspace/src/idp/framework",
    "psycopg2>=2.9.10",
    "alembic>=1.16.1",
    "fastapi-maintenance>=0.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
